--- config/config.json
@@ -1,20 +1,0 @@
-{
-  "development": {
-    "username": "root",
-    "password": null,
-    "database": "challenge",
-    "host": "localhost",
-    "dialect": "mysql"
-  },
-  "test": {
-    "username": "root",
-    "password": null,
-    "database": "database_test",
-    "host": "localhost",
-    "dialect": "mysql"
-  },
-  "production": {
-    "use_env_variable": "JAWSDB_URL",
-    "dialect": "mysql"
-  }
-}
