+++ public/index.html
@@ -22,7 +22,7 @@
          </form>
       </div>
       <div class="row">
-         <form action="/newaccount" method="POST" class="col s12">
+         <form action="/login/account" method="POST" class="col s12">
             <label for="user_name">User Name</label>
             <input id="user_name" type="text" name="username">
             <label for="user_pass">Password</label>
